change: remove dependency on Catcher
Catcher is a great library for error catching, but we do not need it now. Since it collects too much device information and we only used it as a console logger before. (because Flutter always has a bunch of errors, it is not so nice to show a fullscreen tip for every single error!)