fix: (1) when filtering posts by a specified user, the post page should not refresh itself; (2) the status description and actual logic of the "show this user" are inconsistent
(1) This causes the page to become blank when users want to filter by posters only in the deeper levels of the thread.

(2) When a user filters by the poster of a certain post, long-pressing on another post displays "Show all posts," but clicking it actually filters by the user of that other post.